+++ internal/provider/data_source_iosxe_bgp.go
@@ -87,6 +87,14 @@ func (d *BGPDataSource) Schema(ctx context.Context, req datasource.SchemaRequest
 				MarkdownDescription: "Manually configured router identifier",
 				Computed:            true,
 			},
+			"bgp_graceful_restart": schema.BoolAttribute{
+				MarkdownDescription: "Graceful restart capability parameters",
+				Computed:            true,
+			},
+			"bgp_update_delay": schema.Int64Attribute{
+				MarkdownDescription: "Set the max initial delay for sending update",
+				Computed:            true,
+			},
 		},
 	}
 }